+++ src/generator_gemm_amx_microkernel.h
@@ -87,10 +87,10 @@ void libxsmm_generator_gemm_amx_microkernel( libxsmm_generated_code*            
                                                      const libxsmm_gemm_descriptor*     i_xgemm_desc,
                                                      libxsmm_blocking_info_t*           n_blocking_info,
                                                      libxsmm_blocking_info_t*           m_blocking_info,
-                                                     unsigned int                       offset_A,
-                                                     unsigned int                       offset_B,
+                                                     unsigned long long                 offset_A,
+                                                     unsigned long long                 offset_B,
                                                      unsigned int                       is_last_k,
-                                                     int                                i_brgemm_loop,
+                                                     long long                          i_brgemm_loop,
                                                      unsigned int                       fully_unrolled_brloop  );
 
 LIBXSMM_API_INTERN
@@ -101,8 +101,8 @@ void libxsmm_generator_gemm_amx_kernel_kloop( libxsmm_generated_code*           
                                                       const libxsmm_gemm_descriptor*     i_xgemm_desc,
                                                       libxsmm_blocking_info_t*           n_blocking_info,
                                                       libxsmm_blocking_info_t*           m_blocking_info,
-                                                      unsigned int                       A_offs,
-                                                      unsigned int                       B_offs,
+                                                      unsigned long long                 A_offs,
+                                                      unsigned long long                 B_offs,
                                                       unsigned int                       fully_unrolled_brloop );
 
 #endif /* GENERATOR_GEMM_AMX_MICROKERNEL_H */
